[KFI] Finalizing API part 2.
