--- src/Search/OrderDirection.cs
@@ -1,8 +1,0 @@
-﻿namespace SenseNet.Search
-{
-    public enum OrderDirection
-    {
-        Asc,
-        Desc
-    }
-}
