+++ src/Storage/Storage.csproj
@@ -283,7 +283,7 @@
     <Compile Include="Events\NodeOperationEventArgs.cs" />
     <Compile Include="IContentList.cs" />
     <Compile Include="Search\DateTimeParser.cs" />
-    <Compile Include="Search\ISearchEngine.cs" />
+    <Compile Include="Search\Interfaces.cs" />
     <Compile Include="Search\Internal\Operator.cs" />
     <Compile Include="Search\ISearchEngineSupport.cs" />
     <Compile Include="Search\SnPermissionFilterFactory.cs" />
