--- src/Tests/SenseNet.Search.Tests/Implementations/SnQueryLegoBricks.cs
@@ -1,9 +1,0 @@
-﻿using SenseNet.Search.Parser;
-
-namespace SenseNet.Search.Tests.Implementations
-{
-    public class SnQueryLegoBricks
-    {
-        public IPermissionFilterFactory PermissionFilterFactory;
-    }
-}
